pyproject.toml
[project]
name = "ai-judicial-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.203.1",
    "easyocr>=1.7.2",
    "fpdf2>=2.8.4",
    "langchain>=1.0.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.0",
    "langchain-text-splitters>=1.0.0",
    "markdown-pdf>=1.10",
    "pillow>=11.3.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "unstructured[all,docx,pdf]>=0.18.15",
]
